added basic caching for trips;

+++ src/climbingAPI/app/routers/admin.py
@@ -1,14 +1,6 @@
-from typing import List
-
-from fastapi import APIRouter, Body, Depends, HTTPException, Security, status
-from fastapi.encoders import jsonable_encoder
-from fastapi.responses import Response, JSONResponse
+from fastapi import APIRouter, Depends, Security
 from fastapi_auth0 import Auth0User
-from bson import ObjectId
 
-from app.models.spot.spot_model import SpotModel
-from app.models.spot.create_spot_model import CreateSpotModel
-from app.models.spot.update_spot_model import UpdateSpotModel
 from app.core.db import get_db
 from app.core.auth import auth
 
@@ -29,29 +21,29 @@ async def delete_all(user: Auth0User = Security(auth.get_user, scopes=["write:di
 @router.delete('/trips', description="Delete all trips from all users", dependencies=[Depends(auth.implicit_scheme)])
 async def delete_trips(user: Auth0User = Security(auth.get_user, scopes=["write:diary"])):
     db = await get_db()
-    delete_result = await db["trip"].delete_many({})
+    delete_result = await db["trip"].delete_many({"user_id": user.id})
 
 
 @router.delete('/spots', description="Delete all spots from all users", dependencies=[Depends(auth.implicit_scheme)])
 async def delete_spots(user: Auth0User = Security(auth.get_user, scopes=["write:diary"])):
     db = await get_db()
-    delete_result = await db["spot"].delete_many({})
+    delete_result = await db["spot"].delete_many({"user_id": user.id})
 
 
 @router.delete('/routes', description="Delete all routes from all users", dependencies=[Depends(auth.implicit_scheme)])
 async def delete_routes(user: Auth0User = Security(auth.get_user, scopes=["write:diary"])):
     db = await get_db()
-    delete_result = await db["single_pitch_route"].delete_many({})
-    delete_result = await db["multi_pitch_route"].delete_many({})
+    delete_result = await db["single_pitch_route"].delete_many({"user_id": user.id})
+    delete_result = await db["multi_pitch_route"].delete_many({"user_id": user.id})
 
 
 @router.delete('/pitches', description="Delete all pitches from all users", dependencies=[Depends(auth.implicit_scheme)])
 async def delete_pitches(user: Auth0User = Security(auth.get_user, scopes=["write:diary"])):
     db = await get_db()
-    delete_result = await db["pitch"].delete_many({})
+    delete_result = await db["pitch"].delete_many({"user_id": user.id})
 
 
 @router.delete('/ascents', description="Delete all ascents from all users", dependencies=[Depends(auth.implicit_scheme)])
 async def delete_ascents(user: Auth0User = Security(auth.get_user, scopes=["write:diary"])):
     db = await get_db()
-    delete_result = await db["ascent"].delete_many({})
+    delete_result = await db["ascent"].delete_many({"user_id": user.id})
